Conexion BDD 3
select condicion, interfaces

+++ ConexionBDD/src/modelo/articulo.java
@@ -12,8 +12,9 @@ package modelo;
 public class articulo {
     String nombre;
     String descripcion;
-    double precio;
-
+    ////Es recomendable usar tipo float para utilizar decimales
+//    double precio;
+    float precio;
     
     public String getNombre() {
         return nombre;
@@ -31,11 +32,11 @@ public class articulo {
         this.descripcion = descripcion;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(float precio) {
         this.precio = precio;
     }
